Fixed forgotten renaming: 'Registry' to 'DAORegistry'
