+++ packages/subgraph/manifest/data/rinkeby.json
@@ -2,8 +2,8 @@
   "info": "# Do not edit subgraph.yaml,this is a generated file. \n# Instead, edit subgraph.placeholder.yaml and run: yarn manifest",
   "network": "rinkeby",
   "dataSources": {
-    "Registry": {
-      "name": "Registry",
+    "DAORegistry": {
+      "name": "DAORegistry",
       "address": "0xCC5FD930dcbA825e9686F987A9243036Db4A317D",
       "startBlock": 10981816
     }
