+++ packages/subgraph/src/registry.ts
@@ -1,7 +1,6 @@
-import {DAORegistered} from '../generated/Registry/Registry';
+import {DAORegistered} from '../generated/DAORegistry/DAORegistry';
 import {DaoTemplate} from '../generated/templates';
 import {Dao} from '../generated/schema';
-import {handleERC20Token} from './utils/tokens';
 
 export function handleDAORegistered(event: DAORegistered): void {
   let id = event.params.dao.toHexString(); // use dao address as id, because it should not repeat
